+++ src/components/Dropdown/Dropdown.js
@@ -1,71 +1,143 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import classNames from 'classnames'
-import wrapWithClickout from 'react-clickout'
-import SvgSymbol from '../SvgSymbol/SvgSymbol'
-import { ExternalContext } from '../External/External'
+import React, { useState, useEffect, useRef, useCallback } from "react";
+import ReactDOM from "react-dom";
+import PropTypes from "prop-types";
+import classNames from "classnames";
+import { usePopper } from "react-popper";
+import UseEventListener from "../../hooks/UseEventListener";
 
-class Dropdown extends Component {
-  static contextType = ExternalContext
+const Portal = ({ children, querySelector = "#dropdown" }) => {
+  return ReactDOM.createPortal(children, document.querySelector(querySelector));
+};
 
-  state = {
-    isVisible: false,
-  }
+const Dropdown = ({
+  dropdownButton,
+  dropdownContent,
+  className,
+  rootProps,
+  innerClassName,
+  fixedMenu,
+  toggleVisible,
+  isVisible,
+  placement,
+}) => {
+  const [active, setActive] = useState(false);
+  const [visible, setVisible] = useState(false);
+  const referenceRef = useRef();
+  const popperRef = useRef();
 
-  toggleDropdownVisible = () => {
-    this.setState({isVisible: !this.state.isVisible})
-  }
+  const toggle = useCallback(
+    (bool) => {
+      setActive(bool);
+      toggleVisible();
+      setTimeout(() => setVisible(bool), 1);
+    },
+    [toggleVisible]
+  );
 
-  closeDropdown = () => {
-    this.setState({isVisible: false})
-  }
-
-  handleClickout() {
-    if (!this.context.clickoutSuspended) {
-      this.closeDropdown()
+  const { styles, attributes, forceUpdate } = usePopper(
+    referenceRef.current,
+    popperRef.current,
+    {
+      placement: placement || "bottom-end",
+      modifiers: [
+        {
+          name: "preventOverflow",
+          options: {
+            rootBoundary: "viewport",
+            offset: [0, 10],
+          },
+        },
+      ],
     }
-  }
+  );
 
-  render() {
-    const isDropdownVisible =
-      this.props.toggleVisible ? this.props.isVisible : this.state.isVisible
+  useEffect(() => {
+    if (active && forceUpdate) {
+      forceUpdate();
+    }
+  }, [active, forceUpdate]);
 
-    const renderFuncArgs = {
-      isDropdownVisible,
-      toggleDropdownVisible: this.toggleDropdownVisible,
-      closeDropdown: this.closeDropdown
+  useEffect(() => {
+    if (isVisible !== undefined) {
+      if (isVisible && !active) {
+        toggle(true);
+      }
+
+      if (!isVisible && active) {
+        toggle(false);
+      }
+    }
+  }, [isVisible, active, toggle]);
+
+  const handleDocumentClick = (event) => {
+    if (referenceRef.current.contains(event.target)) {
+      return null;
     }
 
-    return (
-      <div className={classNames('mr-dropdown', this.props.className)} {...this.props.rootProps}>
-        {this.props.dropdownButton(renderFuncArgs)}
-        {isDropdownVisible && (
-          <div className={classNames("mr-dropdown__wrapper", this.props.wrapperClassName)}>
-            <div className="mr-dropdown__main">
-              <div className={classNames("mr-dropdown__inner", this.props.innerClassName, {"mr-fixed": this.props.fixedMenu})}>
-                {!this.props.suppressControls &&
-                 <SvgSymbol
-                   sym="icon-triangle"
-                   viewBox="0 0 15 10"
-                   className={classNames("mr-dropdown__arrow", this.props.arrowClassName)}
-                   aria-hidden
-                 />
-                }
-                <div className="mr-dropdown__content">
-                  {this.props.dropdownContent(renderFuncArgs)}
+    if (popperRef.current.contains(event.target)) {
+      return null;
+    }
+
+    toggle(false);
+  };
+
+  UseEventListener("mousedown", handleDocumentClick);
+
+  const renderFuncArgs = {
+    isDropdownVisible: active,
+    toggleDropdownVisible: () => toggle(!active),
+    closeDropdown: () => toggle(false),
+  };
+
+  return (
+    <div data-testid="mr-dropdown" {...rootProps}>
+      <div ref={referenceRef} className={classNames("mr-dropdown", className)}>
+        {dropdownButton(renderFuncArgs)}
+      </div>
+      <Portal>
+        <div
+          ref={popperRef}
+          className="p-0.5 mr-z-250"
+          style={styles.popper}
+          {...attributes.popper}
+        >
+          {active && (
+            <div style={{ visibility: visible ? "visible" : "hidden" }}>
+              <div className="mr-dropdown__main">
+                <div
+                  className={classNames("mr-dropdown__inner", innerClassName, {
+                    "mr-fixed": fixedMenu,
+                  })}
+                >
+                  <div className="mr-dropdown__content">
+                    {dropdownContent(renderFuncArgs)}
+                  </div>
                 </div>
               </div>
             </div>
-          </div>
-        )}
-      </div>
-    )
-  }
-}
+          )}
+        </div>
+      </Portal>
+    </div>
+  );
+};
 
 Dropdown.propTypes = {
   dropdownButton: PropTypes.func.isRequired,
   dropdownContent: PropTypes.func.isRequired,
-}
+  className: PropTypes.string,
+  rootProps: PropTypes.object,
+  innerClassName: PropTypes.string,
+  fixedMenu: PropTypes.bool,
+  suppressControls: PropTypes.bool,
+  arrowClassName: PropTypes.string,
+  toggleVisible: PropTypes.func,
+  isVisible: PropTypes.bool,
+  placement: PropTypes.string,
+};
 
-export default wrapWithClickout(Dropdown)
+Dropdown.defaultProps = {
+  toggleVisible: () => null,
+};
+
+export default Dropdown;